tweaked shaders for unity 5 and cleaned up shaders folder

+++ scatterer/Shaders/customDepthTexture.shader
@@ -21,8 +21,9 @@ v2f vert (appdata_base v) {
     return o;
 }
  
-half4 frag(v2f i) : COLOR {
-    return i.depth.x/i.depth.y;
+float4 frag(v2f i) : COLOR
+{
+    return (i.depth.x/i.depth.y);
 }
 ENDCG
     }
